Add a real Bitmap_Cell oracle and narrow its sF32 locals
tools/test_bitmap.mjs claimed a "Cell source oracle" at line 474 and said op
475's digest "pins the released behavior". No Cell oracle existed: the only
"cell" tokens under tools/oracles/ are local variables in the Bricks fixture
of bitmap_misc_oracle.cpp. Every Cell expectation was a digest of the port's
own output, so any divergence in the rejection loop, the max-- compaction,
the aspect flipxy swap, the per-tile lower-bound cull, the tie-break order or
the 16x16 stride would have kept passing.

bitmap_cell_oracle.cpp transcribes genbitmap.cpp:2694 (Bitmap_Cell is itself
asm-free; only the MMX Fade64 and the RNG needed the scalar treatment the
other oracles already use). Against it the port was byte-exact on ops 475 and
481 but differed on op 452 - and a double-precision build of the same oracle
reproduced all three of the port's outputs exactly, isolating the cause:
v0, v1 and aspdiv are sF32 locals (genbitmap.cpp:2704, 2762) that round to
single precision at every assignment, while the port carried doubles through
sqrt, the mode-1 ratio and the sFPow argument.

With those narrowed per notes/PORT_CONVENTIONS.md, ops 475 and 481 match the
oracle byte for byte over all 2.6 MB and op 452 is down to 35 of 131,072
bytes, each one step of `val`. Op 452 is the only fixture whose amp and gamma
reach sFPow, which is hand-written x87 assembly over fyl2x and f2xm1
(_types.cpp:393); its last ulp is reproducible from neither powf nor
Math.pow, so that residue is recorded in the test rather than papered over.

+++ src/bitmap.js
@@ -1488,9 +1488,12 @@ import { Random } from './core.js';
     }
     const shiftX = 14 - powerOfTwo(bitmap.width), shiftY = 14 - powerOfTwo(bitmap.height);
     const c0 = packedColor(color1), c1 = packedColor(color0), cb = packedColor(color2);
-    const aspect = Math.pow(2, aspectParameter);
-    const aspectFixed = aspect >= 1 ? trunc(65536 / (aspect * aspect)) : trunc(aspect * aspect * 65536);
-    const divisor = aspect >= 1 ? aspect / 16384 : 1 / (16384 * aspect);
+    // aspect, aspf and aspdiv are all sF32 in the released code.
+    const aspect = f32(Math.pow(2, aspectParameter));
+    const aspectFixed = aspect >= 1
+      ? trunc(f32(65536 / f32(aspect * aspect)))
+      : trunc(f32(f32(aspect * aspect) * 65536));
+    const divisor = aspect >= 1 ? f32(aspect / 16384) : f32(1 / f32(16384 * aspect));
     // Mode 2 overwrites the complete temporary pixel on every iteration.
     // Keep one scratch allocation rather than one per generated pixel.
     const cellColor = mode & 2 ? new Uint16Array(4) : null;
@@ -1513,12 +1516,14 @@ import { Random } from './core.js';
           if (distance < best) { second = best; best = distance; bestIndex = i; }
           else if (distance > best && distance < second) second = distance;
         }
-        let value = Math.sqrt(best) * divisor;
+        // v0 and v1 are sF32 locals, so each assignment rounds to single
+        // precision before the next step reads it (genbitmap.cpp:2881-2890).
+        let value = f32(Math.sqrt(best) * divisor);
         if (mode & 1) {
-          const v1 = Math.sqrt(second) * divisor;
-          value = value + v1 > 0.00001 ? (v1 - value) / (v1 + value) : 0;
+          const v1 = f32(Math.sqrt(second) * divisor);
+          value = f32(value + v1) > 0.00001 ? f32((v1 - value) / (v1 + value)) : 0;
         }
-        let fade = clamp15(Math.pow(value * amplitude, gamma) * 0x8000) * 2;
+        let fade = clamp15(f32(Math.pow(f32(value * amplitude), gamma)) * 0x8000) * 2;
         if (mode & 4) fade = 0x10000 - fade;
         const offset4 = (y * bitmap.width + x) * 4;
         if (mode & 2) {
@@ -1602,12 +1607,13 @@ import { Random } from './core.js';
               } else if (distance > best && distance < second) second = distance;
             }
 
-            let value = Math.sqrt(best) * divisor;
+            // Same sF32 narrowing as the scalar path above.
+            let value = f32(Math.sqrt(best) * divisor);
             if (mode & 1) {
-              const v1 = Math.sqrt(second) * divisor;
-              value = value + v1 > 0.00001 ? (v1 - value) / (v1 + value) : 0;
+              const v1 = f32(Math.sqrt(second) * divisor);
+              value = f32(value + v1) > 0.00001 ? f32((v1 - value) / (v1 + value)) : 0;
             }
-            let fade = clamp15(Math.pow(value * amplitude, gamma) * 0x8000) * 2;
+            let fade = clamp15(f32(Math.pow(f32(value * amplitude), gamma)) * 0x8000) * 2;
             if (mode & 4) fade = 0x10000 - fade;
             const offset4 = ((by + ty) * bitmap.width + bx + tx) * 4;
             if (mode & 2) {
